+++ NEWS.md
@@ -24,7 +24,7 @@
   * `bs_delete_message_for_self()` hides a message for you
   * `bs_leave_convo()` leaves a conversation
   * `bs_get_convo_availability()` retrieves info on if you can chat with another actor
-  
+  * `bs_send_message_batch()` sends a batch of messages to different conversations
 
 # bskyr 0.3.0